API transform solver functions into classes (#63)
