+++ skglm/estimators.py
@@ -4,6 +4,7 @@ import warnings
 import numpy as np
 from scipy.sparse import issparse
 from scipy.special import expit
+from skglm.solvers.prox_newton import ProxNewton
 
 from sklearn.utils.validation import check_is_fitted
 from sklearn.utils import check_array, check_consistent_length
@@ -16,17 +17,15 @@ from sklearn.utils.extmath import softmax
 from sklearn.preprocessing import LabelEncoder
 from sklearn.multiclass import OneVsRestClassifier, check_classification_targets
 
-
 from skglm.utils import compiled_clone
-from skglm.solvers import cd_solver_path, multitask_bcd_solver_path
-from skglm.solvers.cd_solver import cd_solver
-from skglm.solvers.multitask_bcd_solver import multitask_bcd_solver
+from skglm.solvers import AndersonCD, MultiTaskBCD
 from skglm.datafits import Quadratic, Logistic, QuadraticSVC, QuadraticMultiTask
 from skglm.penalties import L1, WeightedL1, L1_plus_L2, MCPenalty, IndicatorBox, L2_1
 
 
-def _glm_fit(X, y, model, datafit, penalty):
+def _glm_fit(X, y, model, datafit, penalty, solver):
     is_classif = isinstance(datafit, (Logistic, QuadraticSVC))
+    fit_intercept = solver.fit_intercept
 
     if is_classif:
         check_classification_targets(y)
@@ -45,7 +44,7 @@ def _glm_fit(X, y, model, datafit, penalty):
     else:
         check_X_params = dict(
             dtype=[np.float64, np.float32], order='F',
-            accept_sparse='csc', copy=model.fit_intercept)
+            accept_sparse='csc', copy=fit_intercept)
         check_y_params = dict(ensure_2d=False, order='F')
 
         X, y = model._validate_data(
@@ -68,7 +67,8 @@ def _glm_fit(X, y, model, datafit, penalty):
         raise ValueError("X and y have inconsistent dimensions (%d != %d)"
                          % (n_samples, y.shape[0]))
 
-    if not model.warm_start or not hasattr(model, "coef_"):
+    # if not model.warm_start or not hasattr(model, "coef_"):
+    if not solver.warm_start or not hasattr(model, "coef_"):
         model.coef_ = None
 
     if is_classif and n_classes_ > 2:
@@ -105,23 +105,24 @@ def _glm_fit(X, y, model, datafit, penalty):
     else:
         datafit_jit.initialize(X_, y)
 
-    if model.warm_start and hasattr(model, 'coef_') and model.coef_ is not None:
+    # if model.warm_start and hasattr(model, 'coef_') and model.coef_ is not None:
+    if solver.warm_start and hasattr(model, 'coef_') and model.coef_ is not None:
         if isinstance(datafit, QuadraticSVC):
             w = model.dual_coef_[0, :].copy()
         elif is_classif:
             w = model.coef_[0, :].copy()
         else:
             w = model.coef_.copy()
-        if model.fit_intercept:
+        if fit_intercept:
             w = np.hstack([w, model.intercept_])
-        Xw = X_ @ w[:w.shape[0] - model.fit_intercept] + model.fit_intercept * w[-1]
+        Xw = X_ @ w[:w.shape[0] - fit_intercept] + fit_intercept * w[-1]
     else:
         # TODO this should be solver.get_init() do delegate the work
         if y.ndim == 1:
-            w = np.zeros(n_features + model.fit_intercept, dtype=X_.dtype)
+            w = np.zeros(n_features + fit_intercept, dtype=X_.dtype)
             Xw = np.zeros(n_samples, dtype=X_.dtype)
         else:  # multitask
-            w = np.zeros((n_features + model.fit_intercept, y.shape[1]), dtype=X_.dtype)
+            w = np.zeros((n_features + fit_intercept, y.shape[1]), dtype=X_.dtype)
             Xw = np.zeros(y.shape, dtype=X_.dtype)
 
     # check consistency of weights for WeightedL1
@@ -131,25 +132,12 @@ def _glm_fit(X, y, model, datafit, penalty):
                 "The size of the WeightedL1 penalty weights should be n_features, "
                 "expected %i, got %i." % (X_.shape[1], len(penalty.weights)))
 
-    if is_classif:
-        solver = cd_solver  # TODO to be be replaced by an instance of BaseSolver
-    else:
-        solver = cd_solver if y.ndim == 1 else multitask_bcd_solver
-    # TODO this must be replaced by an instance of BaseSolver being passed
-    # so that arguments are attributes of the `solver` object and arguments
-    # do not need to match across solvers
-    # TODO QUESTIONS
-    # What about ws_strategy?
-    coefs, p_obj, kkt = solver(
-        X_, y, datafit_jit, penalty_jit, w, Xw, max_iter=model.max_iter,
-        max_epochs=model.max_epochs, p0=model.p0,
-        tol=model.tol, fit_intercept=model.fit_intercept,
-        verbose=model.verbose)
+    coefs, p_obj, kkt = solver.solve(X_, y, datafit_jit, penalty_jit, w, Xw)
     model.coef_, model.stop_crit_ = coefs[:n_features], kkt
     if y.ndim == 1:
-        model.intercept_ = coefs[-1] if model.fit_intercept else 0.
+        model.intercept_ = coefs[-1] if fit_intercept else 0.
     else:
-        model.intercept_ = coefs[-1, :] if model.fit_intercept else np.zeros(
+        model.intercept_ = coefs[-1, :] if fit_intercept else np.zeros(
             y.shape[1])
 
     model.n_iter_ = len(p_obj)
@@ -183,30 +171,8 @@ class GeneralizedLinearEstimator(LinearModel):
         Penalty. If None, `penalty` is initialized as a `L1` penalty.
         `penalty` is replaced by a JIT-compiled instance when calling fit.
 
-    max_iter : int, optional
-        The maximum number of iterations (subproblem definitions).
-
-    max_epochs : int
-        Maximum number of CD epochs on each subproblem.
-
-    p0 : int
-        First working set size.
-
-    tol : float, optional
-        Stopping criterion for the optimization.
-
-    fit_intercept : bool, optional (default=True)
-        Whether or not to fit an intercept.
-
-    warm_start : bool, optional (default=False)
-        When set to True, reuse the solution of the previous call to fit as
-        initialization, otherwise, just erase the previous solution.
-
-    ws_strategy : str
-        The score used to build the working set. Can be ``fixpoint`` or ``subdiff``.
-
-    verbose : bool or int
-        Amount of verbosity.
+    solver : instance of BaseSolver, optional
+        Solver. If None, `solver` is initialized as an `AndersonCD` solver.
 
     Attributes
     ----------
@@ -223,20 +189,11 @@ class GeneralizedLinearEstimator(LinearModel):
         Number of subproblems solved to reach the specified tolerance.
     """
 
-    def __init__(self, datafit=None, penalty=None, max_iter=100,
-                 max_epochs=50_000, p0=10, tol=1e-4, fit_intercept=True,
-                 warm_start=False, ws_strategy="subdiff", verbose=0):
+    def __init__(self, datafit=None, penalty=None, solver=None):
         super(GeneralizedLinearEstimator, self).__init__()
-        self.tol = tol
-        self.max_iter = max_iter
-        self.fit_intercept = fit_intercept
-        self.warm_start = warm_start
-        self.verbose = verbose
-        self.max_epochs = max_epochs
-        self.p0 = p0
-        self.ws_strategy = ws_strategy
         self.penalty = penalty
         self.datafit = datafit
+        self.solver = solver
 
     def __repr__(self):
         """Get string representation of the estimator.
@@ -278,7 +235,9 @@ class GeneralizedLinearEstimator(LinearModel):
         """
         self.penalty = self.penalty if self.penalty else L1(1.)
         self.datafit = self.datafit if self.datafit else Quadratic()
-        return _glm_fit(X, y, self, self.datafit, self.penalty)
+        self.solver = self.solver if self.solver else AndersonCD()
+
+        return _glm_fit(X, y, self, self.datafit, self.penalty, self.solver)
 
     def predict(self, X):
         """Predict target values for samples in X.
@@ -386,19 +345,18 @@ class Lasso(LinearModel, RegressorMixin):
     MCPRegression : Sparser regularization than L1 norm.
     """
 
-    def __init__(self, alpha=1., max_iter=100, max_epochs=50_000, p0=10,
-                 verbose=0, tol=1e-4, fit_intercept=True,
-                 warm_start=False, ws_strategy="subdiff"):
+    def __init__(self, alpha=1., max_iter=50, max_epochs=50_000, p0=10, verbose=0,
+                 tol=1e-4, fit_intercept=True, warm_start=False, ws_strategy="subdiff"):
         super().__init__()
+        self.alpha = alpha
         self.tol = tol
         self.max_iter = max_iter
-        self.fit_intercept = fit_intercept
-        self.warm_start = warm_start
-        self.verbose = verbose
         self.max_epochs = max_epochs
         self.p0 = p0
         self.ws_strategy = ws_strategy
-        self.alpha = alpha
+        self.fit_intercept = fit_intercept
+        self.warm_start = warm_start
+        self.verbose = verbose
 
     def fit(self, X, y):
         """Fit the model according to the given training data.
@@ -416,7 +374,12 @@ class Lasso(LinearModel, RegressorMixin):
         self :
             Fitted estimator.
         """
-        return _glm_fit(X, y, self, Quadratic(), L1(self.alpha))
+        # TODO: Add Gram solver
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return _glm_fit(X, y, self, Quadratic(), L1(self.alpha), solver)
 
     def path(self, X, y, alphas, coef_init=None, return_n_iter=True, **params):
         """Compute Lasso path.
@@ -457,13 +420,11 @@ class Lasso(LinearModel, RegressorMixin):
         """
         penalty = compiled_clone(L1(self.alpha))
         datafit = compiled_clone(Quadratic(), to_float32=X.dtype == np.float32)
-
-        return cd_solver_path(
-            X, y, datafit, penalty, alphas=alphas,
-            coef_init=coef_init, max_iter=self.max_iter,
-            return_n_iter=return_n_iter, max_epochs=self.max_epochs,
-            p0=self.p0, tol=self.tol, verbose=self.verbose,
-            ws_strategy=self.ws_strategy)
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return solver.path(X, y, datafit, penalty, alphas, coef_init, return_n_iter)
 
 
 class WeightedLasso(LinearModel, RegressorMixin):
@@ -531,20 +492,20 @@ class WeightedLasso(LinearModel, RegressorMixin):
     Supports weights equal to 0, i.e. unpenalized features.
     """
 
-    def __init__(self, alpha=1., weights=None, max_iter=100, max_epochs=50_000, p0=10,
+    def __init__(self, alpha=1., weights=None, max_iter=50, max_epochs=50_000, p0=10,
                  verbose=0, tol=1e-4, fit_intercept=True, warm_start=False,
                  ws_strategy="subdiff"):
         super().__init__()
+        self.alpha = alpha
+        self.weights = weights
         self.tol = tol
         self.max_iter = max_iter
-        self.fit_intercept = fit_intercept
-        self.warm_start = warm_start
-        self.verbose = verbose
         self.max_epochs = max_epochs
         self.p0 = p0
         self.ws_strategy = ws_strategy
-        self.alpha = alpha
-        self.weights = weights
+        self.fit_intercept = fit_intercept
+        self.warm_start = warm_start
+        self.verbose = verbose
 
     def path(self, X, y, alphas, coef_init=None, return_n_iter=True, **params):
         """Compute Weighted Lasso path.
@@ -588,15 +549,13 @@ class WeightedLasso(LinearModel, RegressorMixin):
             raise ValueError("The number of weights must match the number of \
                               features. Got %s, expected %s." % (
                 len(weights), X.shape[1]))
-
         penalty = compiled_clone(WeightedL1(self.alpha, weights))
         datafit = compiled_clone(Quadratic(), to_float32=X.dtype == np.float32)
-
-        return cd_solver_path(
-            X, y, datafit, penalty, alphas=alphas, coef_init=coef_init,
-            max_iter=self.max_iter, return_n_iter=return_n_iter,
-            max_epochs=self.max_epochs, p0=self.p0, tol=self.tol,
-            verbose=self.verbose)
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return solver.path(X, y, datafit, penalty, alphas, coef_init, return_n_iter)
 
     def fit(self, X, y):
         """Fit the model according to the given training data.
@@ -615,12 +574,15 @@ class WeightedLasso(LinearModel, RegressorMixin):
             Fitted estimator.
         """
         if self.weights is None:
-            warnings.warn(
-                'Weights are not provided, fitting with Lasso penalty')
+            warnings.warn('Weights are not provided, fitting with Lasso penalty')
             penalty = L1(self.alpha)
         else:
             penalty = WeightedL1(self.alpha, self.weights)
-        return _glm_fit(X, y, self, Quadratic(), penalty)
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return _glm_fit(X, y, self, Quadratic(), penalty, solver)
 
 
 class ElasticNet(LinearModel, RegressorMixin):
@@ -643,13 +605,16 @@ class ElasticNet(LinearModel, RegressorMixin):
         combination of L1 and L2.
 
     max_iter : int, optional
-        Maximum number of iterations (subproblem definitions).
+        The maximum number of iterations (subproblem definitions).
 
     max_epochs : int
         Maximum number of CD epochs on each subproblem.
 
     p0 : int
         First working set size.
+
+    verbose : bool or int
+        Amount of verbosity.
 
     tol : float, optional
         Stopping criterion for the optimization.
@@ -661,12 +626,8 @@ class ElasticNet(LinearModel, RegressorMixin):
         When set to True, reuse the solution of the previous call to fit as
         initialization, otherwise, just erase the previous solution.
 
-    verbose : bool or int
-        Amount of verbosity.
-
     ws_strategy : str
-        The score used to build the working set.
-        Can be ``fixpoint`` or ``subdiff``.
+        The score used to build the working set. Can be ``fixpoint`` or ``subdiff``.
 
     Attributes
     ----------
@@ -687,20 +648,20 @@ class ElasticNet(LinearModel, RegressorMixin):
     Lasso : Lasso regularization.
     """
 
-    def __init__(self, alpha=1., l1_ratio=0.5, max_iter=100,
-                 max_epochs=50_000, p0=10, tol=1e-4, fit_intercept=True,
-                 warm_start=False, verbose=0, ws_strategy="subdiff"):
+    def __init__(self, alpha=1., l1_ratio=0.5, max_iter=50, max_epochs=50_000, p0=10,
+                 verbose=0, tol=1e-4, fit_intercept=True, warm_start=False,
+                 ws_strategy="subdiff"):
         super().__init__()
+        self.alpha = alpha
+        self.l1_ratio = l1_ratio
         self.tol = tol
         self.max_iter = max_iter
-        self.fit_intercept = fit_intercept
-        self.warm_start = warm_start
-        self.verbose = verbose
         self.max_epochs = max_epochs
         self.p0 = p0
         self.ws_strategy = ws_strategy
-        self.alpha = alpha
-        self.l1_ratio = l1_ratio
+        self.fit_intercept = fit_intercept
+        self.warm_start = warm_start
+        self.verbose = verbose
 
     def path(self, X, y, alphas, coef_init=None, return_n_iter=True, **params):
         """Compute Elastic Net path.
@@ -741,12 +702,11 @@ class ElasticNet(LinearModel, RegressorMixin):
         """
         penalty = compiled_clone(L1_plus_L2(self.alpha, self.l1_ratio))
         datafit = compiled_clone(Quadratic(), to_float32=X.dtype == np.float32)
-
-        return cd_solver_path(
-            X, y, datafit, penalty, alphas=alphas, coef_init=coef_init,
-            max_iter=self.max_iter, return_n_iter=return_n_iter,
-            max_epochs=self.max_epochs, p0=self.p0, tol=self.tol,
-            verbose=self.verbose)
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return solver.path(X, y, datafit, penalty, alphas, coef_init, return_n_iter)
 
     def fit(self, X, y):
         """Fit the model according to the given training data.
@@ -764,8 +724,12 @@ class ElasticNet(LinearModel, RegressorMixin):
         self :
             Fitted estimator.
         """
-        return _glm_fit(
-            X, y, self, Quadratic(), L1_plus_L2(self.alpha, self.l1_ratio))
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return _glm_fit(X, y, self, Quadratic(),
+                        L1_plus_L2(self.alpha, self.l1_ratio), solver)
 
 
 class MCPRegression(LinearModel, RegressorMixin):
@@ -793,7 +757,7 @@ class MCPRegression(LinearModel, RegressorMixin):
         Should be larger than (or equal to) 1.
 
     max_iter : int, optional
-        Maximum number of iterations (subproblem definitions).
+        The maximum number of iterations (subproblem definitions).
 
     max_epochs : int
         Maximum number of CD epochs on each subproblem.
@@ -836,20 +800,20 @@ class MCPRegression(LinearModel, RegressorMixin):
     Lasso : Lasso regularization.
     """
 
-    def __init__(self, alpha=1., gamma=3, max_iter=100, max_epochs=50_000, p0=10,
+    def __init__(self, alpha=1., gamma=3, max_iter=50, max_epochs=50_000, p0=10,
                  verbose=0, tol=1e-4, fit_intercept=True, warm_start=False,
                  ws_strategy="subdiff"):
         super().__init__()
+        self.alpha = alpha
+        self.gamma = gamma
         self.tol = tol
         self.max_iter = max_iter
-        self.fit_intercept = fit_intercept
-        self.warm_start = warm_start
-        self.verbose = verbose
         self.max_epochs = max_epochs
         self.p0 = p0
         self.ws_strategy = ws_strategy
-        self.alpha = alpha
-        self.gamma = gamma
+        self.fit_intercept = fit_intercept
+        self.warm_start = warm_start
+        self.verbose = verbose
 
     def path(self, X, y, alphas, coef_init=None, return_n_iter=True, **params):
         """Compute MCPRegression path.
@@ -890,12 +854,11 @@ class MCPRegression(LinearModel, RegressorMixin):
         """
         penalty = compiled_clone(MCPenalty(self.alpha, self.gamma))
         datafit = compiled_clone(Quadratic(), to_float32=X.dtype == np.float32)
-
-        return cd_solver_path(
-            X, y, datafit, penalty, alphas=alphas, coef_init=coef_init,
-            max_iter=self.max_iter, return_n_iter=return_n_iter,
-            max_epochs=self.max_epochs, p0=self.p0, tol=self.tol,
-            verbose=self.verbose)
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return solver.path(X, y, datafit, penalty, alphas, coef_init, return_n_iter)
 
     def fit(self, X, y):
         """Fit the model according to the given training data.
@@ -913,8 +876,12 @@ class MCPRegression(LinearModel, RegressorMixin):
         self :
             Fitted estimator.
         """
-        return _glm_fit(
-            X, y, self, Quadratic(), MCPenalty(self.alpha, self.gamma))
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return _glm_fit(X, y, self, Quadratic(), MCPenalty(self.alpha, self.gamma),
+                        solver)
 
 
 class SparseLogisticRegression(LinearClassifierMixin, SparseCoefMixin, BaseEstimator):
@@ -930,32 +897,23 @@ class SparseLogisticRegression(LinearClassifierMixin, SparseCoefMixin, BaseEstim
         Regularization strength; must be a positive float.
 
     tol : float, optional
-        Stopping criterion for the optimization: the solver runs until the
-        duality gap is smaller than ``tol * len(y) * log(2)`` or the
-        maximum number of iteration is reached.
-
-    fit_intercept : bool, optional (default=False)
-        Whether or not to fit an intercept. Currently True is not supported.
+        Stopping criterion for the optimization.
 
     max_iter : int, optional
-        The maximum number of iterations (subproblem definitions).
+        The maximum number of outer iterations (subproblem definitions).
+
+    max_epochs : int
+        Maximum number of prox Newton iterations on each subproblem.
 
     verbose : bool or int
         Amount of verbosity.
 
-    max_epochs : int
-        Maximum number of CD epochs on each subproblem.
-
-    p0 : int
-        First working set size.
+    fit_intercept : bool, optional (default=True)
+        Whether or not to fit an intercept.
 
     warm_start : bool, optional (default=False)
         When set to True, reuse the solution of the previous call to fit as
         initialization, otherwise, just erase the previous solution.
-        Only False is supported so far.
-
-    ws_strategy : str
-        The score used to build the working set. Can be ``fixpoint`` or ``subdiff``.
 
     Attributes
     ----------
@@ -974,20 +932,16 @@ class SparseLogisticRegression(LinearClassifierMixin, SparseCoefMixin, BaseEstim
         Number of subproblems solved to reach the specified tolerance.
     """
 
-    def __init__(
-            self, alpha=1.0, tol=1e-4,
-            fit_intercept=False, max_iter=50, verbose=0,
-            max_epochs=50000, p0=10, warm_start=False, ws_strategy="subdiff"):
+    def __init__(self, alpha=1.0, tol=1e-4, max_iter=20, max_epochs=1_000, verbose=0,
+                 fit_intercept=True, warm_start=False):
         super().__init__()
+        self.alpha = alpha
         self.tol = tol
         self.max_iter = max_iter
+        self.max_epochs = max_epochs
+        self.verbose = verbose
         self.fit_intercept = fit_intercept
         self.warm_start = warm_start
-        self.verbose = verbose
-        self.max_epochs = max_epochs
-        self.p0 = p0
-        self.ws_strategy = ws_strategy
-        self.alpha = alpha
 
     def fit(self, X, y):
         """Fit the model according to the given training data.
@@ -1006,55 +960,11 @@ class SparseLogisticRegression(LinearClassifierMixin, SparseCoefMixin, BaseEstim
         self :
             Fitted estimator.
         """
-        return _glm_fit(X, y, self, Logistic(), L1(self.alpha))
-
-    def path(self, X, y, alphas, coef_init=None, return_n_iter=True, **params):
-        """Compute sparse Logistic Regression path.
-
-        Parameters
-        ----------
-        X : array-like, shape (n_samples, n_features)
-            Training data, where n_samples is the number of samples and
-            n_features is the number of features.
-
-        y : array-like, shape (n_samples,)
-            Target vector relative to X.
-
-        alphas : array
-            Values of regularization strengths for which solutions are
-            computed.
-
-        coef_init : array, shape (n_features,), optional
-            Initial value of the coefficients.
-
-        return_n_iter : bool, optional
-            Return number of iterations along the path.
-
-        **params : kwargs
-            All parameters supported by path.
-
-        Returns
-        -------
-        alphas : array, shape (n_alphas,)
-            The alphas along the path where models are computed.
-
-        coefs : array, shape (n_features, n_alphas)
-            Coefficients along the path.
-
-        stop_crit : array, shape (n_alphas,)
-            Value of stopping criterion at convergence along the path.
-
-        n_iters : array, shape (n_alphas,), optional
-            The number of iterations along the path. If return_n_iter is set to `True`.
-        """
-        penalty = compiled_clone(L1(self.alpha))
-        datafit = compiled_clone(Logistic(), to_float32=X.dtype == np.float32)
-
-        return cd_solver_path(
-            X, y, datafit, penalty, alphas=alphas,
-            coef_init=coef_init, max_iter=self.max_iter,
-            return_n_iter=return_n_iter, max_epochs=self.max_epochs,
-            p0=self.p0, tol=self.tol, verbose=self.verbose)
+        solver = ProxNewton(
+            max_iter=self.max_iter, max_pn_iter=self.max_epochs, tol=self.tol,
+            fit_intercept=self.fit_intercept, warm_start=self.warm_start,
+            verbose=self.verbose)
+        return _glm_fit(X, y, self, Logistic(), L1(self.alpha), solver)
 
     def predict_proba(self, X):
         """Probability estimates.
@@ -1143,18 +1053,18 @@ class LinearSVC(LinearClassifierMixin, SparseCoefMixin, BaseEstimator):
     p0 : int
         First working set size.
 
+    verbose : bool or int
+        Amount of verbosity.
+
     tol : float, optional
         Stopping criterion for the optimization.
 
-    fit_intercept : bool, optional
-        Whether or not to fit an intercept. Currently True is not supported.
+    fit_intercept : bool, optional (default=True)
+        Whether or not to fit an intercept.
 
     warm_start : bool, optional (default=False)
         When set to True, reuse the solution of the previous call to fit as
         initialization, otherwise, just erase the previous solution.
-
-    verbose : bool or int
-        Amount of verbosity.
 
     ws_strategy : str
         The score used to build the working set. Can be ``fixpoint`` or ``subdiff``.
@@ -1177,20 +1087,19 @@ class LinearSVC(LinearClassifierMixin, SparseCoefMixin, BaseEstimator):
         Number of subproblems solved to reach the specified tolerance.
     """
 
-    def __init__(
-            self, C=1., max_iter=100, max_epochs=50_000, p0=10, tol=1e-4,
-            fit_intercept=False, warm_start=False, verbose=0, ws_strategy="subdiff"):
-
+    def __init__(self, C=1., max_iter=50, max_epochs=50_000, p0=10,
+                 verbose=0, tol=1e-4, fit_intercept=True, warm_start=False,
+                 ws_strategy="subdiff"):
         super().__init__()
+        self.C = C
         self.tol = tol
         self.max_iter = max_iter
-        self.fit_intercept = fit_intercept
-        self.warm_start = warm_start
-        self.verbose = verbose
         self.max_epochs = max_epochs
         self.p0 = p0
         self.ws_strategy = ws_strategy
-        self.C = C
+        self.fit_intercept = fit_intercept
+        self.warm_start = warm_start
+        self.verbose = verbose
 
     def fit(self, X, y):
         """Fit LinearSVC classifier.
@@ -1208,7 +1117,11 @@ class LinearSVC(LinearClassifierMixin, SparseCoefMixin, BaseEstimator):
         self
             Fitted estimator.
         """
-        return _glm_fit(X, y, self, QuadraticSVC(), IndicatorBox(self.C))
+        solver = AndersonCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=False,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return _glm_fit(X, y, self, QuadraticSVC(), IndicatorBox(self.C), solver)
 
     # TODO add predict_proba for LinearSVC
 
@@ -1227,7 +1140,7 @@ class MultiTaskLasso(MultiTaskLasso_sklearn):
         Regularization strength (constant that multiplies the L21 penalty).
 
     max_iter : int, optional
-        Maximum number of iterations (subproblem definitions).
+        The maximum number of iterations (subproblem definitions).
 
     max_epochs : int
         Maximum number of CD epochs on each subproblem.
@@ -1239,9 +1152,7 @@ class MultiTaskLasso(MultiTaskLasso_sklearn):
         Amount of verbosity.
 
     tol : float, optional
-        Stopping criterion for the optimization: the solver runs until the
-        duality gap is smaller than ``tol * norm(y) ** 2 / len(y)`` or the
-        maximum number of iteration is reached.
+        Stopping criterion for the optimization.
 
     fit_intercept : bool, optional (default=True)
         Whether or not to fit an intercept.
@@ -1249,6 +1160,9 @@ class MultiTaskLasso(MultiTaskLasso_sklearn):
     warm_start : bool, optional (default=False)
         When set to True, reuse the solution of the previous call to fit as
         initialization, otherwise, just erase the previous solution.
+
+    ws_strategy : str
+        The score used to build the working set. Can be ``fixpoint`` or ``subdiff``.
 
     Attributes
     ----------
@@ -1265,17 +1179,17 @@ class MultiTaskLasso(MultiTaskLasso_sklearn):
         Number of subproblems solved by Celer to reach the specified tolerance.
     """
 
-    def __init__(self, alpha=1., max_iter=100,
-                 max_epochs=50000, p0=10, verbose=0, tol=1e-4,
-                 fit_intercept=True, warm_start=False):
+    def __init__(self, alpha=1., max_iter=50, max_epochs=50_000, p0=10,
+                 verbose=0, tol=1e-4, fit_intercept=True, warm_start=False,
+                 ws_strategy="subdiff"):
         super().__init__(
-            alpha=alpha, tol=tol, max_iter=max_iter,
+            alpha=alpha, tol=tol,
             fit_intercept=fit_intercept, warm_start=warm_start)
-        self.verbose = verbose
-        self.max_epochs = max_epochs
+        self.max_iter = max_iter
         self.p0 = p0
-        self.datafit = QuadraticMultiTask()
-        self.penalty = L2_1(alpha)
+        self.ws_strategy = ws_strategy
+        self.max_epochs = max_epochs
+        self.verbose = verbose
 
     def fit(self, X, Y):
         """Fit MultiTaskLasso model.
@@ -1320,20 +1234,23 @@ class MultiTaskLasso(MultiTaskLasso_sklearn):
         if not self.warm_start or not hasattr(self, "coef_"):
             self.coef_ = None
 
-        _, coefs, kkt = self.path(
-            X, Y, alphas=[self.alpha],
-            coef_init=self.coef_, max_iter=self.max_iter,
-            max_epochs=self.max_epochs, p0=self.p0, verbose=self.verbose,
-            tol=self.tol)
+        datafit_jit = compiled_clone(QuadraticMultiTask(), X.dtype == np.float32)
+        penalty_jit = compiled_clone(L2_1(self.alpha), X.dtype == np.float32)
 
-        self.coef_ = coefs[:, :X.shape[1], 0]
-        self.intercept_ = self.fit_intercept * coefs[:, -1, 0]
-        self.stopping_crit = kkt[-1]
-        self.n_iter_ = len(kkt)
+        solver = MultiTaskBCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        W, obj_out, kkt = solver.solve(X, Y, datafit_jit, penalty_jit)
+
+        self.coef_ = W[:X.shape[1], :].T
+        self.intercept_ = self.fit_intercept * W[-1, :]
+        self.stopping_crit = kkt
+        self.n_iter_ = len(obj_out)
 
         return self
 
-    def path(self, X, Y, alphas, coef_init=None, **params):
+    def path(self, X, Y, alphas, coef_init=None, return_n_iter=False, **params):
         """Compute MultitaskLasso path.
 
         Parameters
@@ -1349,6 +1266,9 @@ class MultiTaskLasso(MultiTaskLasso_sklearn):
 
         coef_init : array, shape (n_features,), optional
             If warm_start is enabled, the optimization problem restarts from coef_init.
+
+        return_n_iter : bool
+            Returns the number of iterations along the path.
 
         **params : kwargs
             All parameters supported by path.
@@ -1367,9 +1287,10 @@ class MultiTaskLasso(MultiTaskLasso_sklearn):
         n_iters : array, shape (n_alphas,), optional
             The number of iterations along the path. If return_n_iter is set to `True`.
         """
-        datafit = compiled_clone(self.datafit, to_float32=X.dtype == np.float32)
-        penalty = compiled_clone(self.penalty)
-
-        return multitask_bcd_solver_path(X, Y, datafit, penalty, alphas=alphas,
-                                         coef_init=coef_init,
-                                         fit_intercept=self.fit_intercept, tol=self.tol)
+        datafit = compiled_clone(QuadraticMultiTask(), to_float32=X.dtype == np.float32)
+        penalty = compiled_clone(L2_1(self.alpha))
+        solver = MultiTaskBCD(
+            self.max_iter, self.max_epochs, self.p0, tol=self.tol,
+            ws_strategy=self.ws_strategy, fit_intercept=self.fit_intercept,
+            warm_start=self.warm_start, verbose=self.verbose)
+        return solver.path(X, Y, datafit, penalty, alphas, coef_init, return_n_iter)
