+++ skglm/solvers/__init__.py
@@ -1,2 +1,9 @@
-from .cd_solver import cd_solver_path  # noqa F401
-from .multitask_bcd_solver import multitask_bcd_solver_path  # noqa F401
+from .anderson_cd import AndersonCD
+from .base import BaseSolver
+from .gram_cd import GramCD
+from .group_bcd import GroupBCD
+from .multitask_bcd import MultiTaskBCD
+from .prox_newton import ProxNewton
+
+
+__all__ = [AndersonCD, BaseSolver, GramCD, GroupBCD, MultiTaskBCD, ProxNewton]
