+++ skglm/tests/test_group.py
@@ -6,11 +6,10 @@ from skglm.penalties import L1
 from skglm.datafits import Quadratic
 from skglm.penalties.block_separable import WeightedGroupL2
 from skglm.datafits.group import QuadraticGroup
-from skglm.solvers.group_bcd_solver import group_bcd_solver
-
+from skglm.solvers import GroupBCD
 from skglm.utils import (
-    _alpha_max_group_lasso, grp_converter, make_correlated_data, AndersonAcceleration)
-from skglm.utils import compiled_clone
+    _alpha_max_group_lasso, grp_converter, make_correlated_data, compiled_clone,
+    AndersonAcceleration)
 from celer import GroupLasso, Lasso
 
 
@@ -36,7 +35,7 @@ def test_check_group_compatible():
     X, y = np.random.randn(5, 5), np.random.randn(5)
 
     with np.testing.assert_raises(Exception):
-        group_bcd_solver(X, y, quad_datafit, l1_penalty)
+        GroupBCD().solve(X, y, quad_datafit, l1_penalty)
 
 
 @pytest.mark.parametrize("n_groups, n_features, shuffle",
@@ -65,7 +64,7 @@ def test_alpha_max(n_groups, n_features, shuffle):
     # compile classes
     quad_group = compiled_clone(quad_group, to_float32=X.dtype == np.float32)
     group_penalty = compiled_clone(group_penalty)
-    w = group_bcd_solver(X, y, quad_group, group_penalty, tol=1e-12)[0]
+    w = GroupBCD(tol=1e-12).solve(X, y, quad_group, group_penalty)[0]
 
     np.testing.assert_allclose(norm(w), 0, atol=1e-14)
 
@@ -89,7 +88,7 @@ def test_equivalence_lasso():
     # compile classes
     quad_group = compiled_clone(quad_group, to_float32=X.dtype == np.float32)
     group_penalty = compiled_clone(group_penalty)
-    w = group_bcd_solver(X, y, quad_group, group_penalty, tol=1e-12)[0]
+    w = GroupBCD(tol=1e-12).solve(X, y, quad_group, group_penalty)[0]
 
     celer_lasso = Lasso(
         alpha=alpha, fit_intercept=False, tol=1e-12, weights=weights).fit(X, y)
@@ -118,7 +117,7 @@ def test_vs_celer_grouplasso(n_groups, n_features, shuffle):
     # compile classes
     quad_group = compiled_clone(quad_group, to_float32=X.dtype == np.float32)
     group_penalty = compiled_clone(group_penalty)
-    w = group_bcd_solver(X, y, quad_group, group_penalty, tol=1e-12)[0]
+    w = GroupBCD(tol=1e-12).solve(X, y, quad_group, group_penalty)[0]
 
     model = GroupLasso(groups=groups, alpha=alpha, weights=weights,
                        fit_intercept=False, tol=1e-12)
@@ -152,9 +151,8 @@ def test_intercept_grouplasso():
 
     quad_group = compiled_clone(quad_group, to_float32=X.dtype == np.float32)
     group_penalty = compiled_clone(group_penalty)
-    w = group_bcd_solver(
-        X, y, quad_group, group_penalty, fit_intercept=True, tol=1e-12)[0]
-
+    w = GroupBCD(fit_intercept=True, tol=1e-12).solve(
+        X, y, quad_group, group_penalty)[0]
     model = GroupLasso(groups=groups, alpha=alpha, weights=weights,
                        fit_intercept=True, tol=1e-12).fit(X, y)
