+++ skglm/tests/test_penalties.py
@@ -9,6 +9,7 @@ from skglm.penalties import (
     L1, L1_plus_L2, WeightedL1, MCPenalty, SCAD, IndicatorBox, L0_5, L2_3,
     L2_1, L2_05, BlockMCPenalty, BlockSCAD)
 from skglm import GeneralizedLinearEstimator
+from skglm.solvers import AndersonCD, MultiTaskBCD
 from skglm.utils import make_correlated_data
 
 
@@ -49,7 +50,7 @@ def test_subdiff_diff(penalty):
     est = GeneralizedLinearEstimator(
         datafit=Quadratic(),
         penalty=penalty,
-        tol=tol,
+        solver=AndersonCD(tol=tol)
     ).fit(X, y)
     # assert the stopping criterion is satisfied
     assert_array_less(est.stop_crit_, tol)
@@ -61,10 +62,10 @@ def test_subdiff_diff_block(block_penalty):
     est = GeneralizedLinearEstimator(
         datafit=QuadraticMultiTask(),
         penalty=block_penalty,
-        tol=tol,
+        solver=MultiTaskBCD(tol=tol)
     ).fit(X, Y)
     # assert the stopping criterion is satisfied
-    assert_array_less(est.stop_crit_, est.tol)
+    assert_array_less(est.stop_crit_, est.solver.tol)
 
 
 if __name__ == "__main__":
